+++ packages/cookie-banner/src/headless/index.ts
@@ -43,6 +43,5 @@ export type {
   BannerTCF,
   TCFGVL,
   TCFGVLVendor,
-  TCFVendor,
   VisitorConsent,
 } from "../types";