+++ packages/cookie-banner/src/index.ts
@@ -48,6 +48,5 @@ export type {
   BannerTCF,
   TCFGVL,
   TCFGVLVendor,
-  TCFVendor,
   VisitorConsent,
 } from "./types";